+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-08 11:42:17
-// Commit: fix(core): correct layout constraints and overflow in core
+// Updated at: 2026-04-08 16:02:36
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '192';
-  final String lastMessage = 'fix(core): correct layout constraints and overflow in core';
-  final DateTime timestamp = DateTime.parse('2026-04-08T11:42:17+05:00');
+  final String commitRef = '193';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-04-08T16:02:36+05:00');
 }